+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ### v1.2 (29.09.2023)
+ - Dictionary has now higher priority than DevelopmentLanguageId
  - Dictionary is now alfabetically sorted using [System.Collections.Specialized.OrderedDictionary]
  - Added [B]reak option to ExtendedStringConfirmation
  - Added progress indicator; files without changes are not reported to the console