+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Dictionary is now alfabetically sorted using [System.Collections.Specialized.OrderedDictionary]
  - Added [B]reak option to ExtendedStringConfirmation
  - Added progress indicator; files without changes are not reported to the console
+ - Fixed bug when saving DeepL credentials; .deepl folder created if not exists
  - Other minor changes
 
 ### v1.1 (06.09.2023)